Separated different resource mapping function into their own classes
