+++ gp2gp-translator/src/main/java/uk/nhs/adaptors/pss/translator/mapper/MedicationRequestMappers/MedicationMapper.java
@@ -1,0 +1,95 @@
+package uk.nhs.adaptors.pss.translator.mapper.MedicationRequestMappers;
+
+import static uk.nhs.adaptors.pss.translator.util.ResourceUtil.generateMeta;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
+
+import org.apache.commons.lang3.StringUtils;
+import org.hl7.fhir.dstu3.model.IdType;
+import org.hl7.fhir.dstu3.model.Medication;
+import org.hl7.fhir.dstu3.model.Reference;
+import org.hl7.fhir.dstu3.model.ResourceType;
+import org.hl7.v3.CD;
+import org.hl7.v3.RCMRMT030101UK04Consumable;
+import org.hl7.v3.RCMRMT030101UK04ManufacturedProduct;
+import org.hl7.v3.RCMRMT030101UK04Material;
+import org.hl7.v3.RCMRMT030101UK04MedicationStatement;
+
+import uk.nhs.adaptors.pss.translator.mapper.CodeableConceptMapper;
+
+public class MedicationMapper {
+    private static final String MEDICATION_URL
+        = "https://fhir.nhs.uk/STU3/StructureDefinition/CareConnect-GPC-Medication-1";
+    private static final Map<String, String> MEDICATION_IDS = new HashMap<>();
+
+    private CodeableConceptMapper codeableConceptMapper;
+
+    protected Medication createMedication(RCMRMT030101UK04Consumable consumable) {
+        if (hasManufacturedMaterial(consumable)) {
+            CD code = consumable.getManufacturedProduct().getManufacturedMaterial().getCode();
+
+            Medication medication = new Medication();
+            medication.setId(getMedicationId(code));
+            medication.setMeta(generateMeta(MEDICATION_URL));
+            medication.setCode(codeableConceptMapper.mapToCodeableConcept(code));
+
+            return medication;
+        }
+        return null;
+    }
+
+    private static String getMedicationId(CD code) {
+        StringBuilder keyBuilder = new StringBuilder();
+        if (code.hasCode()) {
+            keyBuilder.append(code.getCode());
+            keyBuilder.append("-");
+        }
+
+        if (code.hasOriginalText()) {
+            keyBuilder.append(code.getOriginalText());
+            keyBuilder.append("-");
+        }
+
+        if (code.hasDisplayName()) {
+            keyBuilder.append(code.getDisplayName());
+            keyBuilder.append("-");
+        }
+
+        var key = keyBuilder.toString();
+        var value = MEDICATION_IDS.getOrDefault(key, StringUtils.EMPTY);
+
+        if (StringUtils.isNotBlank(value)) {
+            return value;
+        } else {
+            var newId = UUID.randomUUID().toString();
+            MEDICATION_IDS.put(key, newId);
+            return newId;
+        }
+    }
+
+    public static Optional<Reference> extractMedicationReference(RCMRMT030101UK04MedicationStatement medicationStatement) {
+        if (medicationStatement.hasConsumable()) {
+            var medicationCode = medicationStatement.getConsumable()
+                .stream()
+                .filter(MedicationMapper::hasManufacturedMaterial)
+                .map(RCMRMT030101UK04Consumable::getManufacturedProduct)
+                .map(RCMRMT030101UK04ManufacturedProduct::getManufacturedMaterial)
+                .map(RCMRMT030101UK04Material::getCode)
+                .findFirst();
+
+            return medicationCode
+                .map(MedicationMapper::getMedicationId)
+                .map(id -> new IdType(ResourceType.Medication.name(), id))
+                .map(Reference::new);
+        }
+        return Optional.empty();
+    }
+    private static boolean hasManufacturedMaterial(RCMRMT030101UK04Consumable consumable) {
+        return consumable.hasManufacturedProduct() && consumable.getManufacturedProduct().hasManufacturedMaterial()
+            && consumable.getManufacturedProduct().getManufacturedMaterial().hasCode();
+    }
+
+}
